Removing ESLint, Adding a Dev Dependancy for Standard Main page for this project suggests that it supports standard - it isn't installed so installing it.
@dmitriz mentioned in a previous commit :

Eslint and the likes are exactly the tasks I am trying to offload to the
bots :)
That way I don't need to bloat my packages, keeping and  maintaining
more
dependencies can become a pain.
Also Eslint specifically had a high profile security incident and being
that popular is attractive target for hackers.

My thoughts were : Using bots to enforce Standard JS Code standards is
fine - but ensuring that code is 'correct' replies upon me pushing the
code to github, only to then find that it fails code standards.  As the
project shows compliance with Standard, and that standard can be used to
check the code conformance I think that adding in standard would be a
good thing to do - it allows me to check my code before uploading.

If this is a bad idea then this commit can be reverted out - but I find
it useful to know that my code is compliant before pushing.